Updates to fix issues with auto-provision process.
There are 2 variables for specifying the Console secret key. One is set during the server init play and the other is set during the auto-provision init play. So, the Console code has a different key than the auto-provision code. By digging into this I’ve found a couple other issues that should be fixed at the same time:

- update nginx config to use auto-provision php-fpm pool on auto-provision server (currently using easypress-admin)
- update php-fpm to load auto-provision pool config on auto-provision server
- use environment variable for Console secret key in php-fpm easypress-admin pool config (for Console proxy code)
- update the Console proxy code to use the environment variable instead of the exposed key like we’re doing in the code on the auto-provision server